Added commit log
The file still does not verify fully upon loading -- I'm still trying to
figure out the best way to manage loading, verifying, and then keeping
the state in memory in the best way. Ideally, the in-memory
representation would only have a single copy of each duplicate-ly
written header.

The commit log was the last data structure needed for the file. The rest
of the implementation should be finishing todo's in the code, as well as
implementing archiving and checkpointing.

The new file-format.txt has a hand-annotated output of the hex
representation of the database from the basic_op test.