Recover now actually recovers
If any of the disk structures fails to load while reading the "current"
root header and all the data it points to, the second header will then
be tried. This provides the recovery for when a commit fails during
write/sync.

There is still a missing validation before it's safe: the grain data for
the most recent commit batch must be validated. Each grain map and the
actual data must be verified to be fully synced before we proceed with
loading.