Multiple strata are now supported
Fixed multiple bugs:

- Edge case in Ranges type when setting a 1-length range.
- Missing resize before read_exact in CommitLog::read_from
- Logic bugs for finding the correct log entry insert position after
  deserialization
- Log pages were not properly chaining -- the previous_offset was not
  being filled in properly due to how the loop was filtering.
- Basins now have their sequence updated properly when a new stratum is
  allocated

There's a temporary change in behavior for grain allocation that now
limits data being written to more than 4 grains. This is not the correct
solution, but it allowed easier testing of having multiple strata of
differing sizes. This code is on my list of things to rewrite, so I'm
not to concerned with it at the moment.