Added basic benchmark
Sediment isn't ready for benchmarking, but I was curious how good/bad
things were at the moment. Happy with the results, but continuing on
development now that I solved the mystery as to how SQLite supposidly
was using F_FULLFSYNC on macOS while still returning commits in
significantly less time than I was able to achieve a single F_FULLFSYNC
in.