Optimize for Forced Unit Access
All data structures now pad their writes out to the full page, which
enables the kernel to not worry about existing bytes on disk when
queuing up a write.

It might be interesting to consider allowing configurable page sizes to
enable Forced Unit Access where 4,096 bytes doesn't satisfy the
requirements.

Currently the data being written to a grain is *not* written in a way
that enables FUA, unless the grain size is at least 4,096 bytes and the
data being written is a multiple of 4,096 bytes in length.

Without enabling direct IO, I'm not confident there's enough performance
gain to add another page cache to enable FUA for smaller grains. I think
it's fine to rely on the kernel page cache for those pages.