Reduce per-grain overhead
Yesterday I realized that the per-grain overhead of the old strategy for
tracking grain states was roughly 24 bytes. This new strategy reduces it
to 10 bytes per grain.

This new strategy no longer enables looking at a grain and determining
*when* it was allocated or whether it has been archived yet. The only
way to determine if a grain is archived is to look at the commit log
(currently). This could be tracked with in-memory state if needed, but
as of right now, this state isn't needed.

I had originally thought that I needed the information to be able to
handle archiving and freeing, but the commit log already contains
everything needed -- to archive, an entry just needs to be added to the
commit log marking the grains as archived. When that entry is
checkpointed, the grain allocations can be cleared.

Much simpler, and much less overhead.

There's one gotcha that needs to be implemented during free: we must
ensure that after a free operation, attempting to load the data for a
grain will fail the CRC check. This ensures we can check that new data
was written for a grain during the last commit on recovery.