Grain allocations start more conservative
This was a lot of trial and error. In the end, I decided to take an approach of preferring to pack early writes into as small of pre-allocations as possible. When each grain map was only 170 grain, it was ok to allocate nearly every size except the larger ones (writing a 4gb value would allocate 4*170 -- yikes!). With grain map page now being 4084 grains, the idea of growing the file by a multiple of 4084 is scary when the file is only 30kb to begin with.

This change now makes sure that a new grain map page will only grow the file if it's under a certain percentage of the total file length.