Grain CRC is now recorded in the log
While contemplating how to ensure an abandoned write can't accidentally
be considered a valid write inside of a new commit, I realized the
simplest solution was to add the grain CRC to the log. Now, when
validating, the CRC is compared against the log, so the likelihood of
such a situation arising is very low.

This commit causes the rollback tests to fail due to a new todo being
added for recovery.