Added iouring backend
I only have one simple benchmark so far, but despite the overhead of the
abstraction, the performance is basically equivalent to std. This gives
me hope that under more strenuous multithreaded benchmarks that maybe
iouring will shine.

This implementation is also dumb in a way that doesn't seem to impact
performance on std: it reopens the files constantly. I was surprised to
learn that reopening files on both Mac and Linux seems to be incredibly
efficient.