Added multithreaded test
This exposed one edge case: when a new grain map page was allocated but
the commit being executed didn't contain any changes within that page,
the page wasn't being initialized on disk. This broke the future commit,
as the new_pages set didn't contain the offset, and loading all 0's from
disk fails the GrainMapPage CRC check.

By manually adding any missing pages to the modified_pages list, this
ensures the GrainMapPage that is newly allocated is initialized on disk.