Better async buffer management, uncommitted reads
For now, I've enabled uncommitted reads for pending commits. This
probably should be a parameter to get, but for Nebari's purposes, it
really needs this to be able to function in the transaction batching
phase. Without it, because the pending writes are moved into the
committer's next batch, there's no WriteSession active that would give
read access to those grains.

There is still isolation for a WriteSession until either enqueue_commit
or commit is called.

In looking at some early profiles in BonsaiDb, I noticed a lot of
freeing was happening in the commit thread because of
AsyncFileWriter::wait. For async writes, the buffers are never reused,
so I've changed the channels to not retrieve the buffer. This moves the
freeing of the buffers to the threads that wrote the data, and makes the
wait phase faster for the commit thread. This really is a
microoptimization, and ultimately, I'm wondering whether the async
writes are worth it when compared against buffer reuse.