Implement CheckpointGuard
The result of commit operations is now a CheckpointGuard. While the
guard is in memory, it prevents any checkpoint operations from
checkpointing the BatchId that the guard is protecting.

This works well for Nebari's requirements, but one caveat of this
checkpoint guarding is that the checkpoint must be re-requested. I'll
likely tweak this behavior to make it a little more usable for anyone
else using this file format.