Optimized grain allocation (and more)
Benchmarking and profiling is tricky when a significant amount of time
is waiting for hardware, because hardware is inconsistent. These
changes, overall, improve performance, but due to the wide variety of
how long fdatasync can take, individual benchmark runs may or may not
show improved performance compared to the benchmark before these
changes.

The biggest impact is an improvement to grain allocation (Refs #4).
Rather than scanning every basin and every stratum, an index of strata
by grain length is maintained. Additionally, when space cannot be found,
the stratum is updated to note that an allocation of that size can't be
found in it, allowing scans to be skipped until a free operation touches
that stratum.

Speaking of free operations -- it turns out that freed grains were never
communicated to the Atlas, so grains wouldn't be reused until the
database was closed and reopened.

This set of changes also removes quite a lot of memory allocation and
initialization by using a new experiment: rebytes. I've been focusing on
how to reduce pauses during the commit phase, and the syscalls for
allocation were quite significant.

Currently all writes are done synchronously internally. The overhead of
switching threads for the writes and synchronizing did not pan out. I
want to keep experimenting with that approach, however, especially with
rebytes in the mix now.