Initial commit of integrated WAL rewrite
This version is inspired by the original design, but splits the data
storage into multiple files rather than one large file. This will allow
easier implementation of disk reclamation. And, since writes to the
primary data store will be done in the background in batches (thanks to
the WAL), it's theoretically no problem if those multi-file writes are
slightly more work to sync than doing them all within one large file.

This implementation is far from working. Currently, data can be written
to the WAL and read while it's still in the WAL. However, all WAL
recovery and checkpointing is currently unimplemented.