WAL checkpointing
This set of changes adds the ability for the WAL's checkpointing process
to move data from the WAL to the main data files. A simple unit test
showing a database being reopened and being able to read the
checkpointed data passes.

There's a lot left to do, and I'm sure there are some bugs in this new
code. But, we now have data being fully inserted into the database.