Added background fsync operations
This set of changes accomplishes two goals:

- fsyncs the directory when new files are created
- Performs all fsync operations with a background thread pool

This allows the checkpointing thread to write each stratum and begin
syncing the file in its own thread while writing the next stratum. Once
the index file is written and queued for syncing, the checkpointer waits
for the pool to finish all sync operations.

This may be overkill, but it feels like the right thing to do. Because
the WAL is what directly affects the end-user performance, the amount of
overhead syncing in-thread vs allowing multiple syncs to happen
simultaneously may be barely worth it. However, if a database is being
written to so quickly that the checkpointing operation is a bottleneck,
it still needs to be as efficient as it can be.