Embracing poisoning
With this rewrite I was going to avoid using parking_lot given the
recent improvements in std's Mutexes. In the process, however, I had to
deal with poisoning. Initially, I just wanted behavior similar to
parking_lot's, but I quickly realized that detecting that a panic
happened while one of the core structures was held in a mutable state is
actually a very important deal for a database being used in a
multi-threaded environment.

This commit removes all of the PoisonError::into_inner calls and
replaces them with actual error handling.