Better state management with transactions
Previously, the TransactionGuard was being held across the WAL commit.
This ensured the state wasn't published before the WAL was written to,
but it meant that we weren't leveraging OkayWAL's batching since no
other transaction could begin until the TransactionGuard was dropped.

This commit largely is focused on adding TransactionState which the
TransactionLock/Guard now protect access to. This structure keeps track
of all transactions that are being written at any given moment, and more
importantly, allows other transactions to see the current metadata
state.

This introduced me discovering another edge case in OkayWAL: being able
to read data that was written by another committer that is currently
waiting to have its data synchronized to disk. OkayWAL now transparently
handles this by blocking until the bytes are available.

With these changes, the benchmarks are almost running reliably, and the
numbers are looking promising so far.