Fixing benchmark cleanup
I feel like I want to add something to timings to help this situation,
but it's sort of a special requirement. In the meantime, this change
allows detecting when the last instance of a shared database is being
dropped, which happens before the benchmark harness's reset() call. This
means we can shutdown the database fully before our remove_dir_all call
happens.