Actually using the commit log caching
Also fixed a logic bug that would trigger occasionally in the benchmarks
(causing a failed to fill buffer error, or sometimes an infinite loop).
Because GrainIds can be reused, the next entry field could refer to an
already reused GrainId. When iterating backwards, we needed to stop once
we hit the currently checkpointed transaction.

Otherwise, we'd get some other non-CommitLogEntry data or we could end
up causing a cycle to be created in the CommitLogEntries leading to an
infinite loop.