Testing and handling a rolled back stratum
The edge case of what to do if a Stratum is discovered that has two
invalid headers, yet it isn't actually used by the database after
rolling back the latest commit, is now handled. This required a bit of
reworking of how the errors were handled when both headers were missing,
as we can't consider it an error until we know that we need the file.

This change required adding some knowledge of the stratum to the index
header. Previously, we just discovered the files, but now we can know
how many files we are supposed to have.