Switched file operations to file-manager
This commit moves the fsyncs module to a new crate, file-manager, which
restores some of the file manager capabilities that Nebari offered. The
long-term goal is that the file-manager crate can house a single unsafe
call that I want to add that both OkayWAL and Sediment will offer an
option to use -- a call to `fcntl(F_BARRIERFSYNC)` instead of sync_data
for Apple devices (optional and off by default).

The other goal is to unify some of the file synchronizing code.
Specifically the directory syncing code should be able to be shared
between OkayWAL and Sediment such that fewer directory syncs will need
to happen. Since OkayWAL and Sediment don't share any other files, the
rest of the fsync code can largely remain separate except fror sharing
the threadpool as possible.

The main benefit of this change so far is that Sediment now has an
in-memory option.